Update README.md
The covid-19 pandemic represents an unprecedented shock with an international 
impact. Temporary school closures negatively impact students‘ academic achievement as 
they disturb the learning process and cause children to experience substantial learning loss. In 
this project, we use a potential pedagogical function (PPF) to simulate the learning process and 
model shocks representing school closures. We calibrate the model of the learning process to 
reproduce the skills distribution of observed learning outcomes in Tunisia using the PISA 
2015 data. We study two possible scenarios in Tunisia. The optimistic one considers a school 
closure of three months. The pessimistic one considers, in addition to that, another closure of 
half of a school year since Tunisia declared the adoption of the groups‘ system teaching in 
2021. Our simulation shows that the first scenario can cause a learning loss of 1.5 school 
years equivalent to an effect size of 0.54 standard deviations as defined by Cohen. The 
second scenario is found to have a larger impact of almost 3.9 school years which translates 
to a loss of 1.4 standard deviations. The PPF function is flexible and can be used to replicate 
shocks in multiple contexts. Future work could explore additional methods to optimize the 
calibration of the function or be implemented to test the efficacy if remediation strategies in 
reducing the learning loss